pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retrico"
version = "0.1.0"
description = "End-to-end Graph RAG using Knowledgator technologies"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "gliner>=0.2.25",
    "neo4j>=5.0",
    "click>=8.0",
    "falkordblite>=0.1"
]

[project.scripts]
retrico = "retrico.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
llm = [
    "openai>=1.0",
]
falkordb = [
    "falkordb>=1.0",
]
linking = [
    "glinker>=0.1.1"
]
pdf = [
    "pdfminer.six>=20221105",
    "pdfplumber>=0.9",
]
[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
